+++ minsu/projects/movie/css/style.css
@@ -46,9 +46,6 @@ nav li {
   text-align: center;
 }
 
-#inputSearch {
-  cursor: pointer;
-}
 
 .poster{
   width: 40rem;
