+++ minsu/projects/movie/index.html
@@ -20,7 +20,6 @@
     </nav>
     <div class="inputBar">
       <input id="inputValue" autocomplete='off' />
-      <button id="inputSearch">검색</button>
     </div>
   </header>
   <main>
