+++ minsu/projects/movie/script/api.js
@@ -13,3 +13,69 @@ export const request = async(search,page) => {
     alert(e.message)
   }
 }
+
+export const fetchData = async(root, searchResult, firstPage = false) => {
+
+  const {result, tabButton, inputValue, page, favorites} = root.state
+  
+  if(inputValue.trim() === '') return
+  
+  searchResult.setState({
+    ...searchResult.state,
+    isLoading: true,
+  })
+
+  if(tabButton === 'favorite'){
+    searchResult.setState({
+      ...searchResult.state,
+      isLoading: false,
+      result: root.state.favorites.filter( movie => movie.Title.toLowerCase().includes(inputValue.toLowerCase()))
+    })
+  }else{
+    const checkPage = firstPage ? 1 : page
+    const resultData = await request(inputValue, page);
+  
+    if(resultData.Response === 'True' && 'Search' in resultData){
+      if(firstPage) window.scrollTo(0, 0)
+  
+      resultData.Search.forEach(item => {
+        item.favorite = false;
+        const isFavorite = favorites.some(favorite => favorite.imdbID === item.imdbID);
+        if (isFavorite) item.favorite = true;
+      })
+  
+      const nextData = checkPage === 1 ? 
+        resultData.Search : 
+        [...result, ...resultData.Search]
+  
+      root.setState({
+        ...root.state,
+        result: nextData,
+        totalCount: resultData.totalResults,
+        page: firstPage ? 1 : page + 1
+      })
+  
+      searchResult.setState({
+        ...searchResult.state,
+        result: nextData,
+        totalCount: resultData.totalResults,
+        isLoading: false,
+      })
+    }else{
+      root.setState({
+        ...root.state,
+        result: [],
+        totalCount: 0,
+        page: 1
+      })
+  
+      searchResult.setState({
+        ...searchResult.state,
+        result: [],
+        totalCount: 0,
+        isLoading: false,
+      })
+    }
+  }
+
+}
